+++ cartridge/confapplier.lua
@@ -377,8 +377,6 @@ local function boot_instance(clusterwide_config)
         'Unexpected state ' .. vars.state
     )
 
-    invalid_format.start_check()
-
     local topology_cfg = clusterwide_config:get_readonly('topology') or {}
     for _, server in pairs(topology_cfg.servers or {}) do
         if server ~= 'expelled' then
@@ -545,11 +543,12 @@ local function boot_instance(clusterwide_config)
     -- This operation may be long
     -- It recovers snapshot
     -- Or bootstraps replication
+    invalid_format.start_check()
     local snap1 = hotreload.snap_fibers()
     box.cfg(box_opts)
     local snap2 = hotreload.snap_fibers()
     hotreload.whitelist_fibers(hotreload.diff(snap1, snap2))
-
+    invalid_format.end_check()
     require('membership.options').SUSPICIOUSNESS = true
 
     local username = cluster_cookie.username()
@@ -611,8 +610,6 @@ local function boot_instance(clusterwide_config)
             remote_control.resume()
         end
     end
-
-    invalid_format.end_check()
 
     -- Box is ready, start listening full-featured iproto protocol
     remote_control.stop()